requirements.txt
tensorflow==2.18.1
tensorflow-quantum==0.7.6
gymnasium
sympy
numpy
functools import reduce
collections
matplotlib.pyplot
cirq.contrib.svg
yfinance
pandas
numpy
datetime
collections